cleaned up some more and container runs correctly again
