+++ stubs/numba/__init__.pyi
@@ -1,5 +1,5 @@
 from typing import Any
-
+ 
 # Basic types for numba
 float32: Any
 float64: Any
